+++ tests/test_accounting_resource.py
@@ -116,7 +116,7 @@ class TestAccountingResources:
         assert client.grand_total_balance == []
 
     @httpretty.activate
-    def test_get_client__not_found_old_error(self):
+    def test_get_client__not_found_error(self):
         client_id = 12345
         url = "{}/accounting/account/{}/users/clients/{}".format(API_BASE_URL, self.account_id, client_id)
         httpretty.register_uri(
@@ -138,26 +138,6 @@ class TestAccountingResources:
                 "object": "client",
                 "value": "12345"
             }]
-
-    @httpretty.activate
-    def test_get_client__not_found_new_error(self):
-        client_id = 12345
-        url = "{}/accounting/account/{}/users/clients/{}".format(API_BASE_URL, self.account_id, client_id)
-        httpretty.register_uri(
-            httpretty.GET,
-            url,
-            body=json.dumps(get_fixture("get_client_response__not_found_new")),
-            status=404
-        )
-        try:
-            self.freshBooksClient.clients.get(self.account_id, client_id)
-        except FreshBooksError as e:
-            assert str(e) == "Client not found."
-            assert e.status_code == 404
-            assert e.error_code == 1012
-            assert e.error_details == [
-                {"field": "userid", "message": "Client not found.", "object": "client", "value": "12345"}
-            ]
 
     @httpretty.activate
     def test_get_client__bad_response(self):